Add storage_provider interface
Now data can be stored on whatever medium is implemented - right now
it's only in_memory, but a stub for file persistance has been added.
I'll have to figure out how to handle additional parameters (like a
directory where the data will be stored)

Also: I'm not really comfortable with operator interface (namely the
chocice of envvars and it's acquisition being embeded in a module),
however from a coupling point of view it makes perfect sense, as the
main module is unaware of storage kinds, only a very small interface.